requirements.txt
absl-py==1.3.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backoff==2.2.1
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
cachetools==5.2.0
certifi==2023.5.7
charset-normalizer==3.2.0
chromadb==0.3.26
click @ file:///D:/bld/click_1688733153995/work
click-plugins==1.1.1
clickhouse-connect==0.6.6
cligj==0.7.2
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
coloredlogs==15.0.1
commonmark==0.9.1
contextlib2==21.6.0
dataclasses-json==0.5.9
debugpy @ file:///C:/ci/debugpy_1637091961445/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
duckdb==0.8.1
exceptiongroup==1.1.2
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
fastapi==0.100.0
Fiona==1.9.1
flatbuffers==22.12.6
frozenlist==1.3.3
gast==0.4.0
geographiclib==2.0
geopandas==0.12.2
geopy==2.3.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.1
h11==0.14.0
hnswlib==0.7.0
httptools==0.6.0
humanfriendly==10.0
idna==3.4
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1688754491823/work
ipykernel @ file:///D:/bld/ipykernel_1655369313836/work
ipython @ file:///D:/bld/ipython_1685727936079/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1687700988094/work
jupyter_core @ file:///D:/bld/jupyter_core_1686775770267/work
keras==2.11.0
langchain==0.0.228
langchainplus-sdk==0.0.20
libclang==14.0.6
lz4==4.3.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
munch==2.5.0
mypy-extensions==1.0.0
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
numexpr==2.8.4
numpy==1.25.1
oauthlib==3.2.2
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
opt-einsum==3.3.0
osmnx==1.3.0
overrides==7.3.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==1.5.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1688739404342/work
posthog==3.0.1
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1688565951714/work
protobuf==4.23.4
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pulsar-client==3.2.0
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pydantic==1.10.11
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1681904169130/work
pypdf==3.12.1
pyproj==3.4.1
pyreadline3==3.4.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/python-dotenv_1677227146928/work
pytz==2023.3
pywin32==305.1
PyYAML==6.0
pyzmq @ file:///C:/b/abs_655zk4a3s8/croot/pyzmq_1686601465034/work
readme-renderer==37.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rsa==4.9
shapely==2.0.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.0
SQLAlchemy==2.0.18
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
tensorflow-object-detection-api==0.1.1
tokenizers==0.13.3
tornado @ file:///D:/bld/tornado_1656937934674/work
tqdm==4.65.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
twine==4.0.2
typing-inspect==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1688315532570/work
urllib3==2.0.3
uvicorn==0.22.0
watchfiles==0.19.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
websockets==11.0.3
yarl==1.9.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1688902909232/work
zstandard==0.21.0
